+++ example/src/widgets/FitnessWidget.tsx
@@ -1,3 +1,4 @@
+'use no memo';
 /* eslint-disable react-native/no-inline-styles */
 import React from 'react';
 import {